refactor: move getWinners and related tests from RacingManager to Carcollection
