+++ src/test/java/racing/domain/CarCollectionTest.java
@@ -1,8 +1,10 @@
 package racing.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.generator.NumberGenerator;
 import racing.generator.SpecificNumberGenerator;
 
 import java.util.ArrayList;
@@ -46,5 +48,40 @@ class CarCollectionTest {
                 new Car(new Name("test1"), new Position(3)),
                 new Car(new Name("test2"), new Position(5))
         )));
+    }
+
+    @Test
+    void nextStateDontGoForwardTest() {
+        //given
+        CarCollection cars = new CarCollection(List.of(
+                new Car(new Name("test1"), new Position(2)),
+                new Car(new Name("test2"), new Position(4))));
+
+        //when
+        CarCollection carCollection = cars.nextState(new SpecificNumberGenerator(0));
+
+        //then
+        assertThat(carCollection).isEqualTo(new CarCollection(List.of(
+                new Car(new Name("test1"), new Position(2)),
+                new Car(new Name("test2"), new Position(4))
+        )));
+    }
+
+    @Test
+    void getWinnersTest() {
+        //given
+        List<Car> cars = new ArrayList<>();
+        cars.add(new Car("test1"));
+        cars.add(new Car(new Name("test2"), new Position(2)));
+        cars.add(new Car(new Name("test3"), new Position(2)));
+        NumberGenerator numberGenerator = new SpecificNumberGenerator(0);
+        CarCollection carCollection = new CarCollection(cars);
+
+        //when
+        List<String> result = carCollection.getWinners();
+
+        //then
+        Assertions.assertThat(result.get(0)).isEqualTo("test2");
+        Assertions.assertThat(result.get(1)).isEqualTo("test3");
     }
 }
