+++ src/test/java/racing/domain/RacingManagerTest.java
@@ -45,22 +45,4 @@ public class RacingManagerTest {
         assertThat(manager.getCars().get(1).getPosition()).isEqualTo(1);
 
     }
-
-    @Test
-    void getWinnersTest() {
-        //given
-        List<Car> cars = new ArrayList<>();
-        cars.add(new Car("test1"));
-        cars.add(new Car(new Name("test2"), new Position(2)));
-        cars.add(new Car(new Name("test3"), new Position(2)));
-        NumberGenerator numberGenerator = new SpecificNumberGenerator(0);
-        RacingManager manager = new RacingManager(cars, numberGenerator);
-
-        //when
-        List<String> result = manager.getWinners();
-
-        //then
-        assertThat(result.get(0)).isEqualTo("test2");
-        assertThat(result.get(1)).isEqualTo("test3");
-    }
 }
